Argumentos recibidos del JSON combiaron de nombre

+++ db/migrate/2_create_registries.rb
@@ -3,8 +3,8 @@ class CreateRegistries < ActiveRecord::Migration[6.0]
     create_table :registries do |t|
       t.string :name
       t.string :educationLevel
-      t.float :latitude
-      t.float :longitude
+      t.integer :latitude
+      t.integer :longitude
       t.references :users, null: false, foreign_key: true
 
       t.timestamps
